+++ R/format_value.R
@@ -6,6 +6,7 @@
 #' @param missing Value by which \code{NA} values are replaced. By default, an empty string (i.e. \code{""}) is returned for \code{NA}.
 #' @param width Minimum width of the returned string. If not \code{NULL} and \code{width} is larger than the string's length, leading whitespaces are added to the string.
 #' @param as_percent Logical, if \code{TRUE}, value is formatted as percentage value.
+#' @param zap_small Logical, if \code{TRUE}, small values are rounded after \code{digits} decimal places. If \code{FALSE}, values with more decimal places than \code{digits} are printed in scientific notation.
 #' @param ... Arguments passed to or from other methods.
 #'
 #'
@@ -27,23 +28,23 @@
 #'
 #' format_value(iris)
 #' @export
-format_value <- function(x, digits = 2, protect_integers = FALSE, missing = "", width = NULL, as_percent = FALSE, ...) {
+format_value <- function(x, digits = 2, protect_integers = FALSE, missing = "", width = NULL, as_percent = FALSE, zap_small = FALSE, ...) {
   UseMethod("format_value")
 }
 
 
 #' @export
-format_value.data.frame <- function(x, digits = 2, protect_integers = FALSE, missing = "", width = NULL, as_percent = FALSE, ...) {
-  as.data.frame(sapply(x, format_value, digits = digits, protect_integers = protect_integers, missing = missing, width = width, as_percent = as_percent, simplify = FALSE))
+format_value.data.frame <- function(x, digits = 2, protect_integers = FALSE, missing = "", width = NULL, as_percent = FALSE, zap_small = FALSE, ...) {
+  as.data.frame(sapply(x, format_value, digits = digits, protect_integers = protect_integers, missing = missing, width = width, as_percent = as_percent, zap_small = zap_small, simplify = FALSE))
 }
 
 
 #' @export
-format_value.numeric <- function(x, digits = 2, protect_integers = FALSE, missing = "", width = NULL, as_percent = FALSE, ...) {
+format_value.numeric <- function(x, digits = 2, protect_integers = FALSE, missing = "", width = NULL, as_percent = FALSE, zap_small = FALSE, ...) {
   if (protect_integers) {
-    out <- .format_value_unless_integer(x, digits = digits, .missing = missing, .width = width, .as_percent = as_percent, ...)
+    out <- .format_value_unless_integer(x, digits = digits, .missing = missing, .width = width, .as_percent = as_percent, .zap_small = zap_small, ...)
   } else {
-    out <- .format_value(x, digits = digits, .missing = missing, .width = width, .as_percent = as_percent, ...)
+    out <- .format_value(x, digits = digits, .missing = missing, .width = width, .as_percent = as_percent, .zap_small = zap_small, ...)
   }
 
   # Deal with negative zeros
@@ -74,9 +75,9 @@ format_value.logical <- format_value.numeric
 
 
 #' @importFrom stats na.omit
-.format_value_unless_integer <- function(x, digits = 2, .missing = "", .width = NULL, .as_percent = FALSE, ...) {
+.format_value_unless_integer <- function(x, digits = 2, .missing = "", .width = NULL, .as_percent = FALSE, .zap_small = FALSE, ...) {
   if (is.numeric(x) && !all(.is.int(stats::na.omit(x)))) {
-    .format_value(x, digits = digits, .missing = .missing, .width = .width, .as_percent = .as_percent)
+    .format_value(x, digits = digits, .missing = .missing, .width = .width, .as_percent = .as_percent, .zap_small = .zap_small)
   } else if (anyNA(x)) {
     .convert_missing(x, .missing)
   } else {
@@ -86,18 +87,18 @@ format_value.logical <- format_value.numeric
 
 
 
-.format_value <- function(x, digits = 2, .missing = "", .width = NULL, .as_percent = FALSE, ...) {
+.format_value <- function(x, digits = 2, .missing = "", .width = NULL, .as_percent = FALSE, .zap_small = FALSE, ...) {
   # proper character NA
   if (is.na(.missing)) .missing <- NA_character_
 
   if (is.numeric(x)) {
     if (isTRUE(.as_percent)) {
-      need_sci <- (abs(100 * x) >= 1e+5 | (log10(abs(100 * x)) < -digits) & log10(abs(100 * x)) < -3) & x != 0
+      need_sci <- (abs(100 * x) >= 1e+5 | (log10(abs(100 * x)) < -digits) & !.zap_small) & x != 0
       x <- ifelse(is.na(x), .missing,
                   ifelse(need_sci, sprintf("%.*e%%", digits, 100 * x),
                          sprintf("%.*f%%", digits, 100 * x)))
     } else {
-      need_sci <- (abs(x) >= 1e+5 | (log10(abs(x)) < -digits)  & log10(abs(x)) < -3) & x != 0
+      need_sci <- (abs(x) >= 1e+5 | (log10(abs(x)) < -digits) & !.zap_small) & x != 0
       x <- ifelse(is.na(x), .missing,
                   ifelse(need_sci, sprintf("%.*e", digits, x),
                          sprintf("%.*f", digits, x)))